lab4/requirements.txt
numba==0.55.1
numpy~=1.21.5
Pillow==9.0.1
matplotlib==3.5.1